upgrade to latest H3 version; use SPI flash for FS media

+++ src/firmware/src/config/pic32mz_w1_curiosity/crypto/crypto.h
@@ -696,7 +696,7 @@ enum {
 
 /* SHA-384 */
 typedef struct CRYPT_SHA384_CTX {
-    long long holder[32];       /* This structure should be large enough to hold the internal 
+    uint64_t holder[32];       /* This structure should be large enough to hold the internal 
                                    representation, the size is checked during initialization*/
 
 } CRYPT_SHA384_CTX;
@@ -826,7 +826,7 @@ enum {
 
 /* SHA-512 */
 typedef struct CRYPT_SHA512_CTX {
-    long long holder[36];       /* This structure should be large enough to hold 
+    uint64_t holder[36];       /* This structure should be large enough to hold 
                                    the internal representation, the size is checked 
                                    during initialization*/
 
@@ -957,7 +957,7 @@ enum {
 
 /* HMAC */
 typedef struct CRYPT_HMAC_CTX {
-    long long holder[96];   /* This structure should be large enough to hold 
+    uint64_t holder[96];   /* This structure should be large enough to hold 
                                the internal representation, the size is checked 
                                during initialization*/
 } CRYPT_HMAC_CTX;
